Feat: set shipping fee

+++ src/main/java/com/hcmute/shopfee/dto/response/GetShippingFeeResponse.java
@@ -4,5 +4,5 @@ import lombok.Data;
 
 @Data
 public class GetShippingFeeResponse {
-    private Long shippingFee;
+    private int shippingFee;
 }
